+++ lib/skeem/skm_empty_list.rb
@@ -2,7 +2,7 @@ require 'singleton'
 require_relative 'skm_element'
 
 module Skeem
-  # From R7RS: The empty list is a special object of its own type. 
+  # From R7RS: The empty list is a special object of its own type.
   # It is not a pair, it has no elements, and its length is zero.
   class SkmEmptyList < SkmElement
     include Singleton
@@ -14,30 +14,38 @@ module Skeem
     def null?
       true
     end
-    
+
     def pair?
       false
     end
-    
+
     def length
       0
     end
-    
+
     def empty?
       true
     end
-    
+
     def verbatim?
       true
     end
-    
+
     def skm_equal?(other)
       equal?(other)
     end
-    
+
     def to_a
       []
-    end    
+    end
+
+    def klone
+      self
+    end
+    
+    def append_list(aList)
+      aList
+    end
 
     def evaluate(_runtime)
       self